requirements.txt
absl-py==1.2.0
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.2
cycler==0.11.0
cymem==2.0.6
Cython==0.29.30
fonttools==4.37.4
future==0.18.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.49.0
h5py==2.10.0
idna==3.4
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
langcodes==3.3.0
llvmlite==0.29.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
murmurhash==1.0.8
numba==0.45.1
numpy==1.17.2
oauthlib==3.2.1
packaging==21.3
pathy==0.6.2
Pillow==9.1.1
preshed==3.0.7
protobuf==4.21.1
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
thinc==8.1.2
threadpoolctl==3.1.0
torchviz==0.0.2
tqdm==4.64.1
tslearn==0.4.1
typer==0.4.2
typing-extensions==4.1.1
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
zipp==3.8.1
